requirements.txt
setuptools>=42
wheel
asgiref>=3.3.1
Django>=3.1.6
pytz>=2021.1
sqlparse>=0.4.1
psycopg2-yugabytedb~=2.9.3.post0
